Modules
Sorted in Modules to reduce Code in main.c and create some sort of clarity. No functions changed, just reaaranged.
